apps/backend/pyproject.toml
[project]
name = "ai-seo-platform-backend"
version = "0.1.0"
description = "Backend scaffold for AI SEO content platform"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "openai>=1.60.0",
  "PyJWT>=2.9.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.32",
  "alembic>=1.13.2",
  "psycopg[binary]>=3.2.1",
  "redis>=5.0.0",
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.3",
  "python-slugify>=8.0.4",
  "uvicorn[standard]>=0.30.0"
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
